pyproject.toml
[tool.poetry]
name = "samplepackage"
version = "0.1.0"
description = ""
authors = ["endjin <endjineers@endjin.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pyspark = "^3.3.2"
delta-spark = "^2.2.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
jupyter = "^1.0.0"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
